LayoutTests/platform/mac/accessibility/aria-radiobutton-text-expected.txt

One
Two
Three
This tests that aria radio buttons use the text inside of them as the title.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
